Fix rare out-of-bounds error
In rare cases, the computed end index of the last aggregation window might
numerically not be exactly the index of the last element in the time series.
As a consequence, the algorithm assumes that a tiny fraction of the
following element should also be considered in that aggregation window, but
the last element has no following element, resulting in an out-of-bounds
access to the time series vector. Therefore, we add an additional index
check.